+++ web/src/frontend/src/digisos/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Route, Switch } from "react-router";
 import { InjectedIntlProps, injectIntl } from "react-intl";
-import Feilside from "../nav-soknad/components/feilmeldinger/Feilside";
+import Feilside from "../nav-soknad/components/feilside/Feilside";
 import TimeoutBox from "../nav-soknad/components/timeoutbox/TimeoutBox";
 
 import Informasjon from "./informasjon";
@@ -26,7 +26,7 @@ class App extends React.Component<InjectedIntlProps, {}> {
 					<Route
 						component={() => (
 							<Feilside
-								tekst={this.props.intl.formatMessage({id: "feilmelding.404"})}
+								tekst={this.props.intl.formatMessage({ id: "feilmelding.404" })}
 								visTilbakeKnapp={true}
 							/>
 						)}
@@ -36,7 +36,7 @@ class App extends React.Component<InjectedIntlProps, {}> {
 					sessionDurationInMinutes={30}
 					showWarningerAfterMinutes={25}
 				/>
-				<AvbrytSoknad/>
+				<AvbrytSoknad />
 				{this.props.children}
 			</div>
 		);
