+++ web/src/frontend/src/nav-soknad/styles/nav-soknad.less
@@ -2,7 +2,7 @@
 @import "./skjema.less";
 @import "./nav-frontend-skjema-override.less";
 
-@import "../components/feilmeldinger/feilside.less";
+@import "../components/feilside/feilside.less";
 @import "../components/infoblokk/infoblokk.less";
 @import "../components/knapperad/knapperad.less";
 @import "../components/labelMedHjelpetekst/labelMedHjelpetekst.less";
